Improve organization of code, change option parsing. * I think the previous option parsing for input/output files was making incorrect use of the `arguments` functionality from commander. I don't love the fact that you have to use flags for input/output files, but it's okay. * No need for anaemic index.js file, move everything from main.js into this file. * Rewrite code in cli.js so that cli operations are testable.
