--- lib/main.js
@@ -1,149 +1,0 @@
-var fs = require('fs');
-var oboe = require('oboe');
-var curry = require('curry');
-var xmlescape = require('xml-escape');
-
-'use strict';
-
-function extractVerticesEdges(inFilePath, outFilePath, vertexNames, edgeNames, options) {
-  options = options || {};
-  return new Promise(function(fulfill, reject) {
-    var outStream = fs.createWriteStream(outFilePath);
-    outStream.write(`<?xml version="1.0" encoding="UTF-8"?>
-<graphml xmlns="http://graphml.graphdrawing.org/xmlns" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://graphml.graphdrawing.org/xmlns http://graphml.graphdrawing.org/xmlns/1.0/graphml.xsd">
-  <graph id="G" edgedefault="directed">`);
-
-    var vinStream = fs.createReadStream(inFilePath);
-    var einStream = fs.createReadStream(inFilePath);
-
-    vinStream.on('error', function(e) {
-      reject(e);
-    });
-    einStream.on('error', function(e) {
-      reject(e);
-    });
-    var vertexCount = 0;
-    var edgeCount = 0;
-    var pruneOrphanEdges = options.pruneOrphanEdges;
-    var prunedEdgeCount = 0;
-    var ridIndices = {};
-    oboe(vinStream)
-      .node({
-        'records.*': function(record) {
-          if (record['@type'] === 'd' && vertexNames.indexOf(record['@class']) !== -1) {
-            outStream.write('\n'+indent4(vertexToXml(record)));
-            vertexCount++;
-            if (pruneOrphanEdges) {
-              ridIndices[record['@rid']] = true;
-            }
-          }
-          return oboe.drop;
-        }
-      })
-      .done(function() {
-        oboe(einStream)
-          .node({
-            'records.*': function(record) {
-              if (record['@type'] === 'd' && edgeNames.indexOf(record['@class']) !== -1) {
-                if (pruneOrphanEdges) {
-                  var rid1 = record['in'];
-                  var rid2 = record['out'];
-                  if (ridIndices[rid1] !== undefined && ridIndices[rid2] !== undefined) {
-                    outStream.write('\n'+indent4(edgeToXml(record)));
-                    edgeCount++;
-                  }
-                  else {
-                    prunedEdgeCount++;
-                  }
-                }
-                else {
-                  outStream.write('\n'+indent4(edgeToXml(record)));
-                  edgeCount++;
-                }
-              }
-              return oboe.drop;
-            }
-          })
-          .done(function() {
-            outStream.write(`
-  </graph>
-</graphml>
-`);
-            outStream.end();
-      })
-      .on('fail', function(e) {
-        reject(e);
-      });
-  }).on('fail', function(e) {
-    reject(e);
-  });
-    outStream.on('finish', function() {
-      fulfill({edgeCount: edgeCount, vertexCount: vertexCount, prunedEdgeCount: prunedEdgeCount});
-    });
-  });
-}
-
-function convertFile(inFilePath, outFilePath, schemaInfo, options) {
-  var vertexNames = schemaInfo.vertexNames;
-  var edgeNames = schemaInfo.edgeNames;
-  return new Promise(function(fulfill, reject) {
-    extractVerticesEdges(inFilePath, outFilePath, vertexNames, edgeNames, options).
-      then(function(result) {
-        fulfill(result);
-    }).catch(function(err) {
-      reject(err);
-    });
-  });
-}
-
-function vertexToXml(v) {
-  var className = v['@class'];
-  var value;
-  // ugly hack, would fail with nested data structures (though I don't believe those are possible with orientdb)
-  var data = [`<data key="labels">:${v['@class']}</data>`].
-    concat(validVertexKeys(v).map(function(k) {
-      value = v[k];
-      if (typeof value === 'string') {
-        value = xmlescape(value);
-      }
-      else if (Array.isArray(value)) {
-        value = value.map(function(av) {
-          if (typeof av === 'string') {
-            return xmlescape(av);
-          }
-          return av;
-        });
-      }
-      return `<data key="${k}">${value}</data>`;
-  })).join('');
-  return `<node id="${v['@rid']}" labels=":${className}">${data}</node>`;
-}
-
-function edgeToXml(e) {
-  return `<edge id="${e['@rid']}" label="${e['@class']}" source="${e['out']}" target="${e['in']}"><data key="label">${e['@class']}</data></edge>`;
-}
-
-function validVertexKeys(v) {
-  var blacklist = [
-    '@class',
-    '@version',
-    '@rid',
-    '@type',
-    '@fieldTypes'
-  ];
-  var r = /(in|out)_[a-zA-Z]+/;
-  return Object.keys(v).filter(function(k) {
-    return blacklist.indexOf(k) === -1 && k.match(r) === null;
-  });
-}
-
-var indent = curry(function(numSpaces, str) {
-  return ' '.repeat(numSpaces) + str;
-});
-
-var indent4 = indent(4);
-
-module.exports = {
-  convertFile: convertFile,
-  extractVerticesEdges: extractVerticesEdges
-};
